Made a component for PortfolioView on top of home page
