+++ PortfolioHub/%ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,0 +1,1 @@
+1308.2407.15042.0
